feat: ThreeGoodThingsSection を作成
